requirements.txt
# Core dependencies
streamlit==1.29.0
openai==1.6.1
python-docx==1.1.0
lxml==4.9.3

# Utilities
python-dotenv==1.0.0
pandas==2.1.4

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Optional but recommended
requests==2.31.0